# eth/tracers: optimize goja buffer conversion