It seems like going directly from a byte array to `Uint8Array` incurs more copying than going through an `ArrayBuffer` which typed arrays use as the underlying buffer.

Testing this on block 14996244. Before:

```console
Duktape prestateTracerLegacy for block 14996244: 8785ms
Goja prestateTracerLegacy for block 14996244: 17213ms
```

After:

```console
Duktape prestateTracerLegacy for block 14996244: 8913ms
Goja prestateTracerLegacy for block 14996244: 8743ms
```

Fixes #25125